backend/requirements.txt
fastapi
uvicorn
py2neo
requests
beautifulsoup4
pyyaml
